+++ contract/client/src/lib.rs
@@ -1,14 +1,15 @@
 #[macro_use]
 extern crate log;
+extern crate grpcio;
 extern crate serde;
 extern crate serde_cbor;
 
 extern crate ekiden_common;
+extern crate ekiden_compute_api;
 extern crate ekiden_consensus_base;
 extern crate ekiden_contract_common;
 extern crate ekiden_enclave_common;
 extern crate ekiden_registry_base;
-extern crate ekiden_rpc_client;
 extern crate ekiden_scheduler_base;
 extern crate ekiden_storage_base;