+++ contract/trusted/src/lib.rs
@@ -11,14 +11,9 @@ extern crate serde_derive;
 extern crate ekiden_common;
 extern crate ekiden_contract_common;
 extern crate ekiden_enclave_trusted;
-extern crate ekiden_rpc_common;
-extern crate ekiden_rpc_trusted;
 
-pub mod batch;
 pub mod dispatcher;
 #[doc(hidden)]
 pub mod ecalls;
-#[doc(hidden)]
-pub mod rpcs;
 #[macro_use]
 pub mod macros;