+++ contract/untrusted/src/enclave.rs
@@ -12,42 +12,11 @@ pub trait EnclaveContract {
     /// Maximum response size (in kilobytes).
     const MAX_RESPONSE_SIZE: usize = 16 * 1024;
 
-    /// Check if the enclave has a batch ready for execution and copy it over.
-    fn contract_take_batch(&self) -> Result<CallBatch>;
-
     /// Invoke a contract on a batch of calls and return the (encrypted) outputs.
     fn contract_call_batch(&self, batch: &CallBatch) -> Result<OutputBatch>;
 }
 
 impl EnclaveContract for Enclave {
-    fn contract_take_batch(&self) -> Result<CallBatch> {
-        // Reserve space up to the maximum size of serialized response.
-        let mut response: Vec<u8> = Vec::with_capacity(Self::MAX_RESPONSE_SIZE * 1024);
-        let mut response_length = 0;
-
-        let status = unsafe {
-            ecall_proxy::contract_take_batch(
-                self.get_id(),
-                response.as_mut_ptr() as *mut u8,
-                response.capacity(),
-                &mut response_length,
-            )
-        };
-
-        if status != sgx_status_t::SGX_SUCCESS {
-            return Err(Error::new(format!(
-                "contract_take_batch: failed to call enclave ({})",
-                status
-            )));
-        }
-
-        unsafe {
-            response.set_len(response_length);
-        }
-
-        Ok(serde_cbor::from_slice(&response)?)
-    }
-
     fn contract_call_batch(&self, batch: &CallBatch) -> Result<OutputBatch> {
         // Encode input batch.
         let batch_encoded = serde_cbor::to_vec(batch)?;