+++ CHANGELOG.md
@@ -43,6 +43,7 @@
 * Added the multilayer storage backend described in RFC 0004.
 * You can now configure the last resort layer in the multilayer storage backend.
 * Handle stragglers in dummy consensus backend.
+* Change async contract call submission protocol to avoid the use of enclave RPC.
 
 # 0.1.0